[REFACTOR] #89 링크 목록 화면(좋아요, 안읽음) ViewModel 리펙토링
