+++ feature/linklist/src/main/java/pokitmons/pokit/linklist/Preview.kt
@@ -5,7 +5,11 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import pokitmons.pokit.core.feature.model.paging.PagingState
 import pokitmons.pokit.core.ui.theme.PokitTheme
+import pokitmons.pokit.linklist.model.Link
 import pokitmons.pokit.linklist.model.LinkListScreenState
 
 @Preview(showBackground = true)
@@ -16,18 +20,40 @@ fun Preview() {
             modifier = Modifier.fillMaxSize()
         ) {
             LinkListScreen(
-                state = LinkListScreenState(),
+                viewModel = dummyLinkListViewModel,
                 onBackPressed = { },
-                loadNextLinkList = { },
-                toggleSort = {},
-                showLinkDetailBottomSheet = {},
-                hideLinkDetailBottomSheet = {},
-                showCheckLinkRemoveBottomSheet = {},
-                hideCheckLinkRemoveBottomSheet = {},
-                onClickBookmark = {},
                 onClickModifyLink = {},
-                onClickLinkRemove = {}
             )
         }
     }
+}
+
+private val dummyLinkListScreenState = LinkListScreenState()
+
+private val dummyLinkListViewModel = object : LinkListViewModel {
+    override val state: StateFlow<LinkListScreenState>
+        get() = MutableStateFlow(dummyLinkListScreenState)
+    override val linkList: StateFlow<List<Link>>
+        get() = MutableStateFlow(emptyList())
+    override val linkListState: StateFlow<PagingState>
+        get() = MutableStateFlow(PagingState.IDLE)
+
+    override fun loadNextLinks() {}
+
+    override fun refreshLinks() {}
+
+    override fun toggleSortType() {}
+
+    override fun toggleBookmark(link: Link) {}
+
+    override fun showLinkDetailBottomSheet(link: Link) {}
+
+    override fun hideLinkDetailBottomSheet() {}
+
+    override fun showCheckLinkRemoveBottomSheet() {}
+
+    override fun hideCheckLinkRemoveBottomSheet() {}
+
+    override fun removeLink(linkId: String) {}
+
 }
